src/micropython/modopenearable/status_led.cpp: fixed warnings

+++ src/micropython/modopenearable/status_led.cpp
@@ -2,7 +2,7 @@
 #include "StateIndicator.h"
 
 mp_obj_t openearable_set_led_color(mp_obj_t r, mp_obj_t g, mp_obj_t b) {
-    RGBColor color = {mp_obj_get_int(r), mp_obj_get_int(g), mp_obj_get_int(b)};
+    RGBColor color = {(uint8_t) mp_obj_get_int(r), (uint8_t) mp_obj_get_int(g), (uint8_t) mp_obj_get_int(b)};
     state_indicator.set_custom_color(color);
     return mp_const_none;
 }
